refactor(MathematicalExpressionParser) : int 타입보다 큰 수에 대응하기 위해 long type으로 변경
- MathematicalExpressionParser.parse() : long[] 반환하도록 변경
- BasicMathematicalExpressionParser : Long.parseLong()으로 변경 및 변수들 long type으로 변경

+++ src/main/java/calculator/parser/BasicMathematicalExpressionParser.java
@@ -2,9 +2,9 @@ package calculator.parser;
 
 public class BasicMathematicalExpressionParser implements MathematicalExpressionParser {
 
-	private int stringNumberToNumber(String stringNumber) {
+	private long stringNumberToNumber(String stringNumber) {
 		try {
-			int number = Integer.parseInt(stringNumber);
+			long number = Long.parseLong(stringNumber);
 			if (number <= 0) {
 				throw new IllegalArgumentException(ParserError.INVALID_POSITIVE_NUMBER.getMessage());
 			}
@@ -16,9 +16,9 @@ public class BasicMathematicalExpressionParser implements MathematicalExpression
 	}
 
 	@Override
-	public int[] parse(String text, String separatorRegex) {
+	public long[] parse(String text, String separatorRegex) {
 		String[] stringNumbers = text.split(separatorRegex);
-		int[] numbers = new int[stringNumbers.length];
+		long[] numbers = new long[stringNumbers.length];
 
 		for (int i = 0; i < stringNumbers.length; i++) {
 			numbers[i] = stringNumberToNumber(stringNumbers[i]);
